criado delete de contas

+++ src/business/contas/ContasBusiness.ts
@@ -9,52 +9,65 @@ export default class ContasBusiness {
     readonly objetivosFinanceirosRepository = new ObjetivosFinanceirosRepository
     readonly movimentacoesRepository = new MovimentacoesRepository
 
-async index (idUsuario : number){
-  
-    return await this.contasRepository.readConta(idUsuario)
-}
+    async index(idUsuario: number) {
 
-    async cadastrarContas (conta : Contas){
+        return await this.contasRepository.readConta(idUsuario)
+    }
+
+    async cadastrarContas(conta: Contas) {
 
 
-        conta.valorTotal = (Number(conta.valorLivre)  + Number(conta.valorSeparado))
+        conta.valorTotal = (Number(conta.valorLivre) + Number(conta.valorSeparado))
 
         return await this.contasRepository.insertConta(conta)
     }
 
-    async deleteConta(idConta: number){
-        
+    async deleteConta(idConta: number) {
 
-try{
 
-    const  retorno = await this.contasRepository.deleteContaId(idConta)
-    return retorno
+        try {
 
-} catch(e){
+            let retorno
+            let verificar = await this.movimentacoesRepository.verificarMovimentacoesContas(idConta)
+            if (!verificar) {
+                verificar = await this.objetivosFinanceirosRepository.verificarObjetivosFinanceirosContas(idConta)
+                if (!verificar) {
+                    retorno = await this.contasRepository.deleteContaId(idConta)
+                } else {
+                    retorno = "Existe Objetivos Ja cadastrados."
+                }
 
-}
+            } else {
+                retorno = "Existe Movimentaçôes já cadastrados."
+            }
+
+            return retorno
+
+        } catch (e) {
+            return e
+        }
 
     }
-    
-    async updateConta(contas : Contas ){
-      
-        const  verificarContas = await this.contasRepository.buscarSaldoContasRpository(contas.id)
-  
-if(!contas.valorLivre){
-    contas.valorLivre = verificarContas.valorLivre
-}
 
-        
-if(!contas.valorSeparado){
-    contas.valorSeparado = verificarContas.valorSeparado
-}
+    async updateConta(contas: Contas) {
 
-console.log(contas)
+        const verificarContas = await this.contasRepository.buscarSaldoContasRpository(contas.id)
 
-//const  retornoContas = await this.contasRepository.updateContas(contas)
+        if (!contas.valorLivre) {
+            contas.valorLivre = verificarContas.valorLivre
+        }
 
 
-        
+        if (!contas.valorSeparado) {
+            contas.valorSeparado = verificarContas.valorSeparado
+        }
+
+        console.log(contas)
+
+        //const  retornoContas = await this.contasRepository.updateContas(contas)
+
+
+
     }
